ai_qa_agent/requirements.txt
streamlit>=1.31.0
appium-python-client>=4.4.0
selenium>=4.17.0
pyyaml>=6.0.1
requests>=2.31.0

